[spaces] several fixes for pre release
- update after review
